Crate google-books1

This documentation was generated from books crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

bookshelves

get, list and volumes.list

cloudloading

add book, delete book and update book

dictionary

list offline metadata

layers

annotation data.get, annotation data.list, get, list, volume annotations.get and volume annotations.list

myconfig

get user settings, release download access, request access, sync volume licenses and update user settings

mylibrary

annotations.delete, annotations.insert, annotations.list, annotations.summary, annotations.update, bookshelves.add volume, bookshelves.clear volumes, bookshelves.get, bookshelves.list, bookshelves.move volume, bookshelves.remove volume, bookshelves.volumes.list, readingpositions.get and readingpositions.set position

onboarding

list categories and list category volumes

promooffer

accept, dismiss and get

volumes

associated.list, get, list, mybooks.list, recommended.list, recommended.rate and useruploaded.list

Everything else about the books v1 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.volumes().useruploaded_list(...).doit()
let r = hub.myconfig().sync_volume_licenses(...).doit()
let r = hub.volumes().list(...).doit()
let r = hub.volumes().associated_list(...).doit()
let r = hub.bookshelves().volumes_list(...).doit()
let r = hub.volumes().recommended_list(...).doit()
let r = hub.mylibrary().bookshelves_volumes_list(...).doit()
let r = hub.volumes().mybooks_list(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-books1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-books1" as books1;
use books1::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.volumes().list("q")
             .start_index(11)
             .source("dolores")
             .show_preorders(false)
             .projection("sadipscing")
             .print_type("aliquyam")
             .partner("ea")
             .order_by("no")
             .max_results(80)
             .library_restrict("justo")
             .lang_restrict("et")
             .filter("et")
             .download("diam")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Annotation
There is no detailed description.

AnnotationClientVersionRanges
Selection ranges sent from the client.

AnnotationCurrentVersionRanges
Selection ranges for the most recent content version.

AnnotationLayerSummary
There is no detailed description.

Annotationdata
There is no detailed description.

Annotations
There is no detailed description.

AnnotationsSummary
There is no detailed description.

AnnotationsSummaryLayers
There is no detailed description.

Annotationsdata
There is no detailed description.

Books
Central instance to access all Books related resource activities

BooksAnnotationsRange
There is no detailed description.

BooksCloudloadingResource
There is no detailed description.

BooksVolumesRecommendedRateResponse
There is no detailed description.

Bookshelf
There is no detailed description.

BookshelveGetCallBuilder
Retrieves metadata for a specific bookshelf for the specified user.

BookshelveListCallBuilder
Retrieves a list of public bookshelves for the specified user.

BookshelveMethodsBuilder
A builder providing access to all methods supported on bookshelve resources. It is not used directly, but through the Books hub.

BookshelveVolumeListCallBuilder
Retrieves volumes in a specific bookshelf for the specified user.

Bookshelves
There is no detailed description.

Category
There is no detailed description.

CategoryItems
A list of onboarding categories.

CloudloadingAddBookCallBuilder

CloudloadingDeleteBookCallBuilder
Remove the book and its contents

CloudloadingMethodsBuilder
A builder providing access to all methods supported on cloudloading resources. It is not used directly, but through the Books hub.

CloudloadingUpdateBookCallBuilder

ConcurrentAccessRestriction
There is no detailed description.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DictionaryListOfflineMetadataCallBuilder
Returns a list of offline dictionary meatadata available

DictionaryMethodsBuilder
A builder providing access to all methods supported on dictionary resources. It is not used directly, but through the Books hub.

Dictlayerdata
There is no detailed description.

DictlayerdataCommon
There is no detailed description.

DictlayerdataDict
There is no detailed description.

DictlayerdataDictSource
The source, url and attribution for this dictionary data.

DictlayerdataDictWords
There is no detailed description.

DictlayerdataDictWordsDerivatives
There is no detailed description.

DictlayerdataDictWordsDerivativesSource
There is no detailed description.

DictlayerdataDictWordsExamples
There is no detailed description.

DictlayerdataDictWordsExamplesSource
There is no detailed description.

DictlayerdataDictWordsSenses
There is no detailed description.

DictlayerdataDictWordsSensesConjugations
There is no detailed description.

DictlayerdataDictWordsSensesDefinitions
There is no detailed description.

DictlayerdataDictWordsSensesDefinitionsExamples
There is no detailed description.

DictlayerdataDictWordsSensesDefinitionsExamplesSource
There is no detailed description.

DictlayerdataDictWordsSensesSource
There is no detailed description.

DictlayerdataDictWordsSensesSynonyms
There is no detailed description.

DictlayerdataDictWordsSensesSynonymsSource
There is no detailed description.

DictlayerdataDictWordsSource
The words with different meanings but not related words, e.g. "go" (game) and "go" (verb).

DownloadAccessRestriction
There is no detailed description.

DownloadAccesses
There is no detailed description.

Geolayerdata
There is no detailed description.

GeolayerdataCommon
There is no detailed description.

GeolayerdataGeo
There is no detailed description.

GeolayerdataGeoBoundary
There is no detailed description.

GeolayerdataGeoViewport
The viewport for showing this location. This is a latitude, longitude rectangle.

GeolayerdataGeoViewportHi
There is no detailed description.

GeolayerdataGeoViewportLo
There is no detailed description.

LayerAnnotationDataGetCallBuilder
Gets the annotation data.

LayerAnnotationDataListCallBuilder
Gets the annotation data for a volume and layer.

LayerGetCallBuilder
Gets the layer summary for a volume.

LayerListCallBuilder
List the layer summaries for a volume.

LayerMethodsBuilder
A builder providing access to all methods supported on layer resources. It is not used directly, but through the Books hub.

LayerVolumeAnnotationGetCallBuilder
Gets the volume annotation.

LayerVolumeAnnotationListCallBuilder
Gets the volume annotations for a volume and layer.

Layersummaries
There is no detailed description.

Layersummary
There is no detailed description.

Metadata
There is no detailed description.

MetadataItems
A list of offline dictionary metadata.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

MyconfigGetUserSettingCallBuilder
Gets the current settings for the user.

MyconfigMethodsBuilder
A builder providing access to all methods supported on myconfig resources. It is not used directly, but through the Books hub.

MyconfigReleaseDownloadAccesCallBuilder
Release downloaded content access restriction.

MyconfigRequestAccesCallBuilder
Request concurrent and download access restrictions.

MyconfigSyncVolumeLicenseCallBuilder
Request downloaded content access for specified volumes on the My eBooks shelf.

MyconfigUpdateUserSettingCallBuilder
Sets the settings for the user. Unspecified sub-objects will retain the existing value.

MylibraryAnnotationDeleteCallBuilder
Deletes an annotation.

MylibraryAnnotationInsertCallBuilder
Inserts a new annotation.

MylibraryAnnotationListCallBuilder
Retrieves a list of annotations, possibly filtered.

MylibraryAnnotationSummaryCallBuilder
Gets the summary of specified layers.

MylibraryAnnotationUpdateCallBuilder
Updates an existing annotation.

MylibraryBookshelveAddVolumeCallBuilder
Adds a volume to a bookshelf.

MylibraryBookshelveClearVolumeCallBuilder
Clears all volumes from a bookshelf.

MylibraryBookshelveGetCallBuilder
Retrieves metadata for a specific bookshelf belonging to the authenticated user.

MylibraryBookshelveListCallBuilder
Retrieves a list of bookshelves belonging to the authenticated user.

MylibraryBookshelveMoveVolumeCallBuilder
Moves a volume within a bookshelf.

MylibraryBookshelveRemoveVolumeCallBuilder
Removes a volume from a bookshelf.

MylibraryBookshelveVolumeListCallBuilder
Gets volume information for volumes on a bookshelf.

MylibraryMethodsBuilder
A builder providing access to all methods supported on mylibrary resources. It is not used directly, but through the Books hub.

MylibraryReadingpositionGetCallBuilder
Retrieves my reading position information for a volume.

MylibraryReadingpositionSetPositionCallBuilder
Sets my reading position information for a volume.

Offers
There is no detailed description.

OffersItems
A list of offers.

OffersItemsItems
There is no detailed description.

OnboardingListCategoryCallBuilder
List categories for onboarding experience.

OnboardingListCategoryVolumeCallBuilder
List available volumes under categories for onboarding experience.

OnboardingMethodsBuilder
A builder providing access to all methods supported on onboarding resources. It is not used directly, but through the Books hub.

PromoofferAcceptCallBuilder

PromoofferDismisCallBuilder

PromoofferGetCallBuilder
Returns a list of promo offers available to the user

PromoofferMethodsBuilder
A builder providing access to all methods supported on promooffer resources. It is not used directly, but through the Books hub.

ReadingPosition
There is no detailed description.

RequestAccess
There is no detailed description.

Review
There is no detailed description.

ReviewAuthor
Author of this review.

ReviewSource
Information regarding the source of this review, when the review is not from a Google Books user.

Usersettings
There is no detailed description.

UsersettingsNotesExport
User settings in sub-objects, each for different purposes.

Volume
There is no detailed description.

Volume2
There is no detailed description.

VolumeAccessInfo
Any information about a volume related to reading or obtaining that volume text. This information can depend on country (books may be public domain in one country but not in another, e.g.).

VolumeAccessInfoEpub
Information about epub content. (In LITE projection.)

VolumeAccessInfoPdf
Information about pdf content. (In LITE projection.)

VolumeAssociatedListCallBuilder
Return a list of associated books.

VolumeGetCallBuilder
Gets volume information for a single volume.

VolumeLayerInfo
What layers exist in this volume and high level information about them.

VolumeLayerInfoLayers
A layer should appear here if and only if the layer exists for this book.

VolumeListCallBuilder
Performs a book search.

VolumeMethodsBuilder
A builder providing access to all methods supported on volume resources. It is not used directly, but through the Books hub.

VolumeMybookListCallBuilder
Return a list of books in My Library.

VolumeRecommendedInfo
Recommendation related information for this volume.

VolumeRecommendedListCallBuilder
Return a list of recommended books for the current user.

VolumeRecommendedRateCallBuilder
Rate a recommended book for the current user.

VolumeSaleInfo
Any information about a volume related to the eBookstore and/or purchaseability. This information can depend on the country where the request originates from (i.e. books may not be for sale in certain countries).

VolumeSaleInfoListPrice
Suggested retail price. (In LITE projection.)

VolumeSaleInfoOffers
Offers available for this volume (sales and rentals).

VolumeSaleInfoOffersListPrice
Offer list (=undiscounted) price in Micros.

VolumeSaleInfoOffersRentalDuration
The rental duration (for rental offers only).

VolumeSaleInfoOffersRetailPrice
Offer retail (=discounted) price in Micros

VolumeSaleInfoRetailPrice
The actual selling price of the book. This is the same as the suggested retail or list price unless there are offers or discounts on this volume. (In LITE projection.)

VolumeSearchInfo
Search result information related to this volume.

VolumeUserInfo
User specific information related to this volume. (e.g. page this user last read or whether they purchased this book)

VolumeUserInfoCopy
Copy/Paste accounting information.

VolumeUserInfoRentalPeriod
Period during this book is/was a valid rental.

VolumeUserInfoUserUploadedVolumeInfo
There is no detailed description.

VolumeUseruploadedListCallBuilder
Return a list of books uploaded by the current user.

VolumeVolumeInfo
General volume information.

VolumeVolumeInfoDimensions
Physical dimensions of this volume.

VolumeVolumeInfoImageLinks
A list of image links for all the sizes that are available. (In LITE projection.)

VolumeVolumeInfoIndustryIdentifiers
Industry standard identifiers for this volume.

Volumeannotation
There is no detailed description.

VolumeannotationContentRanges
The content ranges to identify the selected text.

Volumeannotations
There is no detailed description.

Volumes
There is no detailed description.

Enums

Result
A universal result type used as return for all action method results.

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

ResponseResult
Identifies types which are used in API responses.

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.