Module google_books1::api

Structs

Annotation
There is no detailed description.

AnnotationClientVersionRanges
Selection ranges sent from the client.

AnnotationCurrentVersionRanges
Selection ranges for the most recent content version.

AnnotationLayerSummary
There is no detailed description.

Annotations
There is no detailed description.

AnnotationsSummary
There is no detailed description.

AnnotationsSummaryLayers
There is no detailed description.

Annotationsdata
There is no detailed description.

Books
Central instance to access all Books related resource activities

BooksAnnotationsRange
There is no detailed description.

BooksCloudloadingResource
There is no detailed description.

BooksVolumesRecommendedRateResponse
There is no detailed description.

Bookshelf
There is no detailed description.

BookshelveGetCall
Retrieves metadata for a specific bookshelf for the specified user.

BookshelveListCall
Retrieves a list of public bookshelves for the specified user.

BookshelveMethods
A builder providing access to all methods supported on bookshelve resources. It is not used directly, but through the Books hub.

BookshelveVolumeListCall
Retrieves volumes in a specific bookshelf for the specified user.

Bookshelves
There is no detailed description.

Category
There is no detailed description.

CategoryItems
A list of onboarding categories.

CloudloadingAddBookCall
Add a user-upload volume and triggers processing.

CloudloadingDeleteBookCall
Remove the book and its contents

CloudloadingMethods
A builder providing access to all methods supported on cloudloading resources. It is not used directly, but through the Books hub.

CloudloadingUpdateBookCall
Updates a user-upload volume.

ConcurrentAccessRestriction
There is no detailed description.

DictionaryAnnotationdata
There is no detailed description.

DictionaryListOfflineMetadataCall
Returns a list of offline dictionary metadata available

DictionaryMethods
A builder providing access to all methods supported on dictionary resources. It is not used directly, but through the Books hub.

Dictlayerdata
There is no detailed description.

DictlayerdataCommon
There is no detailed description.

DictlayerdataDict
There is no detailed description.

DictlayerdataDictSource
The source, url and attribution for this dictionary data.

DictlayerdataDictWords
There is no detailed description.

DictlayerdataDictWordsDerivatives
There is no detailed description.

DictlayerdataDictWordsDerivativesSource
There is no detailed description.

DictlayerdataDictWordsExamples
There is no detailed description.

DictlayerdataDictWordsExamplesSource
There is no detailed description.

DictlayerdataDictWordsSenses
There is no detailed description.

DictlayerdataDictWordsSensesConjugations
There is no detailed description.

DictlayerdataDictWordsSensesDefinitions
There is no detailed description.

DictlayerdataDictWordsSensesDefinitionsExamples
There is no detailed description.

DictlayerdataDictWordsSensesDefinitionsExamplesSource
There is no detailed description.

DictlayerdataDictWordsSensesSource
There is no detailed description.

DictlayerdataDictWordsSensesSynonyms
There is no detailed description.

DictlayerdataDictWordsSensesSynonymsSource
There is no detailed description.

DictlayerdataDictWordsSource
The words with different meanings but not related words, e.g. "go" (game) and "go" (verb).

Discoveryclusters
There is no detailed description.

DiscoveryclustersClusters
There is no detailed description.

DiscoveryclustersClustersBannerWithContentContainer
There is no detailed description.

DownloadAccessRestriction
There is no detailed description.

DownloadAccesses
There is no detailed description.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

FamilyInfo
There is no detailed description.

FamilyInfoMembership
Family membership info of the user that made the request.

FamilysharingGetFamilyInfoCall
Gets information regarding the family that the user is part of.

FamilysharingMethods
A builder providing access to all methods supported on familysharing resources. It is not used directly, but through the Books hub.

FamilysharingShareCall
Initiates sharing of the content with the user's family. Empty response indicates success.

FamilysharingUnshareCall
Initiates revoking content that has already been shared with the user's family. Empty response indicates success.

GeoAnnotationdata
There is no detailed description.

Geolayerdata
There is no detailed description.

GeolayerdataCommon
There is no detailed description.

GeolayerdataGeo
There is no detailed description.

GeolayerdataGeoViewport
The viewport for showing this location. This is a latitude, longitude rectangle.

GeolayerdataGeoViewportHi
There is no detailed description.

GeolayerdataGeoViewportLo
There is no detailed description.

LayerAnnotationDataGetCall
Gets the annotation data.

LayerAnnotationDataListCall
Gets the annotation data for a volume and layer.

LayerGetCall
Gets the layer summary for a volume.

LayerListCall
List the layer summaries for a volume.

LayerMethods
A builder providing access to all methods supported on layer resources. It is not used directly, but through the Books hub.

LayerVolumeAnnotationGetCall
Gets the volume annotation.

LayerVolumeAnnotationListCall
Gets the volume annotations for a volume and layer.

Layersummaries
There is no detailed description.

Layersummary
There is no detailed description.

Metadata
There is no detailed description.

MetadataItems
A list of offline dictionary metadata.

MyconfigGetUserSettingCall
Gets the current settings for the user.

MyconfigMethods
A builder providing access to all methods supported on myconfig resources. It is not used directly, but through the Books hub.

MyconfigReleaseDownloadAccesCall
Release downloaded content access restriction.

MyconfigRequestAccesCall
Request concurrent and download access restrictions.

MyconfigSyncVolumeLicenseCall
Request downloaded content access for specified volumes on the My eBooks shelf.

MyconfigUpdateUserSettingCall
Sets the settings for the user. If a sub-object is specified, it will overwrite the existing sub-object stored in the server. Unspecified sub-objects will retain the existing value.

MylibraryAnnotationDeleteCall
Deletes an annotation.

MylibraryAnnotationInsertCall
Inserts a new annotation.

MylibraryAnnotationListCall
Retrieves a list of annotations, possibly filtered.

MylibraryAnnotationSummaryCall
Gets the summary of specified layers.

MylibraryAnnotationUpdateCall
Updates an existing annotation.

MylibraryBookshelveAddVolumeCall
Adds a volume to a bookshelf.

MylibraryBookshelveClearVolumeCall
Clears all volumes from a bookshelf.

MylibraryBookshelveGetCall
Retrieves metadata for a specific bookshelf belonging to the authenticated user.

MylibraryBookshelveListCall
Retrieves a list of bookshelves belonging to the authenticated user.

MylibraryBookshelveMoveVolumeCall
Moves a volume within a bookshelf.

MylibraryBookshelveRemoveVolumeCall
Removes a volume from a bookshelf.

MylibraryBookshelveVolumeListCall
Gets volume information for volumes on a bookshelf.

MylibraryMethods
A builder providing access to all methods supported on mylibrary resources. It is not used directly, but through the Books hub.

MylibraryReadingpositionGetCall
Retrieves my reading position information for a volume.

MylibraryReadingpositionSetPositionCall
Sets my reading position information for a volume.

Notification
There is no detailed description.

NotificationGetCall
Returns notification details for a given notification id.

NotificationMethods
A builder providing access to all methods supported on notification resources. It is not used directly, but through the Books hub.

Offers
There is no detailed description.

OffersItems
A list of offers.

OffersItemsItems
There is no detailed description.

OnboardingListCategoryCall
List categories for onboarding experience.

OnboardingListCategoryVolumeCall
List available volumes under categories for onboarding experience.

OnboardingMethods
A builder providing access to all methods supported on onboarding resources. It is not used directly, but through the Books hub.

PersonalizedstreamGetCall
Returns a stream of personalized book clusters

PersonalizedstreamMethods
A builder providing access to all methods supported on personalizedstream resources. It is not used directly, but through the Books hub.

PromoofferAcceptCall
Accepts the promo offer.

PromoofferDismisCall
Marks the promo offer as dismissed.

PromoofferGetCall
Returns a list of promo offers available to the user

PromoofferMethods
A builder providing access to all methods supported on promooffer resources. It is not used directly, but through the Books hub.

ReadingPosition
There is no detailed description.

RequestAccessData
There is no detailed description.

Review
There is no detailed description.

ReviewAuthor
Author of this review.

ReviewSource
Information regarding the source of this review, when the review is not from a Google Books user.

Series
There is no detailed description.

SeriesSeries
There is no detailed description.

SeriesSeriesSeriesSubscriptionReleaseInfo
There is no detailed description.

SeriesSeriesSeriesSubscriptionReleaseInfoCurrentReleaseInfo
There is no detailed description.

SeriesSeriesSeriesSubscriptionReleaseInfoNextReleaseInfo
There is no detailed description.

Seriesmembership
There is no detailed description.

SeryGetCall
Returns Series metadata for the given series ids.

SeryMembershipGetCall
Returns Series membership data given the series id.

SeryMethods
A builder providing access to all methods supported on sery resources. It is not used directly, but through the Books hub.

Usersettings
There is no detailed description.

UsersettingsNotesExport
User settings in sub-objects, each for different purposes.

UsersettingsNotification
There is no detailed description.

UsersettingsNotificationMatchMyInterests
There is no detailed description.

UsersettingsNotificationMoreFromAuthors
There is no detailed description.

UsersettingsNotificationMoreFromSeries
There is no detailed description.

UsersettingsNotificationPriceDrop
There is no detailed description.

UsersettingsNotificationRewardExpirations
There is no detailed description.

Volume
There is no detailed description.

Volume2
There is no detailed description.

VolumeAccessInfo
Any information about a volume related to reading or obtaining that volume text. This information can depend on country (books may be public domain in one country but not in another, e.g.).

VolumeAccessInfoEpub
Information about epub content. (In LITE projection.)

VolumeAccessInfoPdf
Information about pdf content. (In LITE projection.)

VolumeAssociatedListCall
Return a list of associated books.

VolumeGetCall
Gets volume information for a single volume.

VolumeLayerInfo
What layers exist in this volume and high level information about them.

VolumeLayerInfoLayers
A layer should appear here if and only if the layer exists for this book.

VolumeListCall
Performs a book search.

VolumeMethods
A builder providing access to all methods supported on volume resources. It is not used directly, but through the Books hub.

VolumeMybookListCall
Return a list of books in My Library.

VolumeRecommendedInfo
Recommendation related information for this volume.

VolumeRecommendedListCall
Return a list of recommended books for the current user.

VolumeRecommendedRateCall
Rate a recommended book for the current user.

VolumeSaleInfo
Any information about a volume related to the eBookstore and/or purchaseability. This information can depend on the country where the request originates from (i.e. books may not be for sale in certain countries).

VolumeSaleInfoListPrice
Suggested retail price. (In LITE projection.)

VolumeSaleInfoOffers
Offers available for this volume (sales and rentals).

VolumeSaleInfoOffersListPrice
Offer list (=undiscounted) price in Micros.

VolumeSaleInfoOffersRentalDuration
The rental duration (for rental offers only).

VolumeSaleInfoOffersRetailPrice
Offer retail (=discounted) price in Micros

VolumeSaleInfoRetailPrice
The actual selling price of the book. This is the same as the suggested retail or list price unless there are offers or discounts on this volume. (In LITE projection.)

VolumeSearchInfo
Search result information related to this volume.

VolumeUserInfo
User specific information related to this volume. (e.g. page this user last read or whether they purchased this book)

VolumeUserInfoCopy
Copy/Paste accounting information.

VolumeUserInfoFamilySharing
Information on the ability to share with the family.

VolumeUserInfoRentalPeriod
Period during this book is/was a valid rental.

VolumeUserInfoUserUploadedVolumeInfo
There is no detailed description.

VolumeUseruploadedListCall
Return a list of books uploaded by the current user.

VolumeVolumeInfo
General volume information.

VolumeVolumeInfoDimensions
Physical dimensions of this volume.

VolumeVolumeInfoImageLinks
A list of image links for all the sizes that are available. (In LITE projection.)

VolumeVolumeInfoIndustryIdentifiers
Industry standard identifiers for this volume.

VolumeVolumeInfoPanelizationSummary
A top-level summary of the panelization info in this volume.

VolumeVolumeInfoReadingModes
The reading modes available for this volume.

Volumeannotation
There is no detailed description.

VolumeannotationContentRanges
The content ranges to identify the selected text.

Volumeannotations
There is no detailed description.

Volumes
There is no detailed description.

Volumeseriesinfo
There is no detailed description.

VolumeseriesinfoVolumeSeries
There is no detailed description.

VolumeseriesinfoVolumeSeriesIssue
List of issues. Applicable only for Collection Edition and Omnibus.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.