Struct google_books1::Annotationdata

pub struct Annotationdata {
    pub annotation_type: Option<String>,
    pub kind: Option<String>,
    pub updated: Option<String>,
    pub volume_id: Option<String>,
    pub encoded_data: Option<String>,
    pub layer_id: Option<String>,
    pub data: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

annotation data get layers (response)

Fields

annotation_type

The type of annotation this data is for.

kind

Resource Type

updated

Timestamp for the last time this data was updated. (RFC 3339 UTC date-time format).

volume_id

The volume id for this data. *

encoded_data

Base64 encoded data for this annotation data.

layer_id

The Layer id for this data. *

data

no description provided

id

Unique id for this annotation data.

self_link

URL for this resource. *