Struct google_books1::Books

pub struct Books<C, A> {
    // some fields omitted
}

Central instance to access all Books related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_books1 as books1;
use books1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.volumes().list("q")
             .start_index(82)
             .source("gubergren")
             .show_preorders(false)
             .projection("aliquyam")
             .print_type("ea")
             .partner("no")
             .order_by("justo")
             .max_results(80)
             .max_allowed_maturity_rating("et")
             .library_restrict("et")
             .lang_restrict("diam")
             .filter("ipsum")
             .download("Lorem")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Methods

impl<'a, C, A> Books<C, A> where C: BorrowMut<Client>, A: GetToken

fn new(client: C, authenticator: A) -> Books<C, A>

fn bookshelves(&'a self) -> BookshelveMethods<'a, C, A>

fn cloudloading(&'a self) -> CloudloadingMethods<'a, C, A>

fn dictionary(&'a self) -> DictionaryMethods<'a, C, A>

fn layers(&'a self) -> LayerMethods<'a, C, A>

fn myconfig(&'a self) -> MyconfigMethods<'a, C, A>

fn mylibrary(&'a self) -> MylibraryMethods<'a, C, A>

fn notification(&'a self) -> NotificationMethods<'a, C, A>

fn onboarding(&'a self) -> OnboardingMethods<'a, C, A>

fn personalizedstream(&'a self) -> PersonalizedstreamMethods<'a, C, A>

fn promooffer(&'a self) -> PromoofferMethods<'a, C, A>

fn series(&'a self) -> SeryMethods<'a, C, A>

fn volumes(&'a self) -> VolumeMethods<'a, C, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.1.15.

Returns the previously set user-agent.