Struct google_books1::CloudloadingAddBookCall

pub struct CloudloadingAddBookCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder for the addBook method supported by a cloudloading resource. It is not used directly, but through a CloudloadingMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.cloudloading().add_book()
             .upload_client_token("et")
             .name("clita")
             .mime_type("eos")
             .drive_document_id("dolores")
             .doit();

Methods

impl<'a, C, A> CloudloadingAddBookCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, BooksCloudloadingResource)>

Perform the operation you have build so far.

fn upload_client_token(self, new_value: &str) -> CloudloadingAddBookCall<'a, C, A>

Sets the upload_client_token query property to the given value.

fn name(self, new_value: &str) -> CloudloadingAddBookCall<'a, C, A>

The document name. It can be set only if the drive_document_id is set.

Sets the name query property to the given value.

fn mime_type(self, new_value: &str) -> CloudloadingAddBookCall<'a, C, A>

The document MIME type. It can be set only if the drive_document_id is set.

Sets the mime_type query property to the given value.

fn drive_document_id(self, new_value: &str) -> CloudloadingAddBookCall<'a, C, A>

A drive document id. The upload_client_token must not be set.

Sets the drive_document_id query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> CloudloadingAddBookCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> CloudloadingAddBookCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> CloudloadingAddBookCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Full.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.