Struct google_books1::ConcurrentAccessRestriction

pub struct ConcurrentAccessRestriction {
    pub nonce: Option<String>,
    pub kind: Option<String>,
    pub restricted: Option<bool>,
    pub volume_id: Option<String>,
    pub max_concurrent_devices: Option<i32>,
    pub device_allowed: Option<bool>,
    pub source: Option<String>,
    pub time_window_seconds: Option<i32>,
    pub signature: Option<String>,
    pub reason_code: Option<String>,
    pub message: Option<String>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

nonce

Client nonce for verification. Download access and client-validation only.

kind

Resource type.

restricted

Whether this volume has any concurrent access restrictions.

volume_id

Identifies the volume for which this entry applies.

max_concurrent_devices

The maximum number of concurrent access licenses for this volume.

device_allowed

Whether access is granted for this (user, device, volume).

source

Client app identifier for verification. Download access and client-validation only.

time_window_seconds

Time in seconds for license auto-expiration.

signature

Response signature.

reason_code

Error/warning reason code.

message

Error/warning message.