Struct google-books1::GeolayerdataGeo

pub struct GeolayerdataGeo {
    pub country_code: String,
    pub zoom: i32,
    pub longitude: f64,
    pub map_type: String,
    pub latitude: f64,
    pub boundary: Vec<Vec<GeolayerdataGeoBoundary>>,
    pub viewport: GeolayerdataGeoViewport,
    pub cache_policy: String,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

country_code

The country code of the location.

zoom

The Zoom level to use for the map. Zoom levels between 0 (the lowest zoom level, in which the entire world can be seen on one map) to 21+ (down to individual buildings). See: https://developers.google.com/maps/documentation/staticmaps/#Zoomlevels

longitude

The longitude of the location.

map_type

The type of map that should be used for this location. EX: HYBRID, ROADMAP, SATELLITE, TERRAIN

latitude

The latitude of the location.

boundary

The boundary of the location as a set of loops containing pairs of latitude, longitude coordinates.

viewport

The viewport for showing this location. This is a latitude, longitude rectangle.

cache_policy

The cache policy active for this data. EX: UNRESTRICTED, RESTRICTED, NEVER