Struct google-books1::LayerMethodsBuilder

pub struct LayerMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on layer resources. It is not used directly, but through the Books hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-books1" as books1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `annotation_data_get(...)`, `annotation_data_list(...)`, `get(...)`, `list(...)`, `volume_annotations_get(...)` and `volume_annotations_list(...)`
// to build up your call.
let rb = hub.layers();

Methods

impl<'a, C, NC, A> LayerMethodsBuilder<'a, C, NC, A>

fn annotation_data_get(&self, volume_id: &str, layer_id: &str, annotation_data_id: &str, content_version: &str) -> LayerAnnotationDataGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the annotation data.

fn volume_annotations_get(&self, volume_id: &str, layer_id: &str, annotation_id: &str) -> LayerVolumeAnnotationGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the volume annotation.

fn list(&self, volume_id: &str) -> LayerListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

List the layer summaries for a volume.

fn get(&self, volume_id: &str, summary_id: &str) -> LayerGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the layer summary for a volume.

fn volume_annotations_list(&self, volume_id: &str, layer_id: &str, content_version: &str) -> LayerVolumeAnnotationListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the volume annotations for a volume and layer.

fn annotation_data_list(&self, volume_id: &str, layer_id: &str, content_version: &str) -> LayerAnnotationDataListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the annotation data for a volume and layer.