Struct google-books1::Layersummary

pub struct Layersummary {
    pub kind: String,
    pub annotation_count: i32,
    pub data_count: i32,
    pub annotation_types: Vec<String>,
    pub updated: String,
    pub volume_id: String,
    pub annotations_data_link: String,
    pub annotations_link: String,
    pub content_version: String,
    pub layer_id: String,
    pub volume_annotations_version: String,
    pub id: String,
    pub self_link: String,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get layers (response)

Fields

kind

Resource Type

annotation_count

The number of annotations for this layer.

data_count

The number of data items for this layer.

annotation_types

The list of annotation types contained for this layer.

updated

Timestamp for the last time an item in this layer was updated. (RFC 3339 UTC date-time format).

volume_id

The volume id this resource is for.

annotations_data_link

Link to get data for this annotation.

annotations_link

The link to get the annotations for this layer.

content_version

The content version this resource is for.

layer_id

The layer id for this summary.

volume_annotations_version

The current version of this layer's volume annotations. Note that this version applies only to the data in the books.layers.volumeAnnotations.* responses. The actual annotation data is versioned separately.

id

Unique id of this layer summary.

self_link

URL to this resource.