Struct google_books1::MyconfigMethods

pub struct MyconfigMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on myconfig resources. It is not used directly, but through the Books hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_books1 as books1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get_user_settings(...)`, `release_download_access(...)`, `request_access(...)`, `sync_volume_licenses(...)` and `update_user_settings(...)`
// to build up your call.
let rb = hub.myconfig();

Methods

impl<'a, C, A> MyconfigMethods<'a, C, A>

fn request_access(&self, source: &str, volume_id: &str, nonce: &str, cpksver: &str) -> MyconfigRequestAccesCall<'a, C, A>

Create a builder to help you perform the following task:

Request concurrent and download access restrictions.

Arguments

source - String to identify the originator of this request.

volumeId - The volume to request concurrent/download restrictions for.

nonce - The client nonce value.

cpksver - The device/version ID from which to request the restrictions.

fn release_download_access(&self, volume_ids: &Vec<String>, cpksver: &str) -> MyconfigReleaseDownloadAccesCall<'a, C, A>

Create a builder to help you perform the following task:

Release downloaded content access restriction.

Arguments

volumeIds - The volume(s) to release restrictions for.

cpksver - The device/version ID from which to release the restriction.

fn sync_volume_licenses(&self, source: &str, nonce: &str, cpksver: &str) -> MyconfigSyncVolumeLicenseCall<'a, C, A>

Create a builder to help you perform the following task:

Request downloaded content access for specified volumes on the My eBooks shelf.

Arguments

source - String to identify the originator of this request.

nonce - The client nonce value.

cpksver - The device/version ID from which to release the restriction.

fn get_user_settings(&self) -> MyconfigGetUserSettingCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the current settings for the user.

fn update_user_settings(&self, request: Usersettings) -> MyconfigUpdateUserSettingCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the settings for the user. If a sub-object is specified, it will overwrite the existing sub-object stored in the server. Unspecified sub-objects will retain the existing value.

Arguments

request - No description provided.