Struct google_books1::MylibraryMethods

pub struct MylibraryMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on mylibrary resources. It is not used directly, but through the Books hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_books1 as books1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `annotations_delete(...)`, `annotations_insert(...)`, `annotations_list(...)`, `annotations_summary(...)`, `annotations_update(...)`, `bookshelves_add_volume(...)`, `bookshelves_clear_volumes(...)`, `bookshelves_get(...)`, `bookshelves_list(...)`, `bookshelves_move_volume(...)`, `bookshelves_remove_volume(...)`, `bookshelves_volumes_list(...)`, `readingpositions_get(...)` and `readingpositions_set_position(...)`
// to build up your call.
let rb = hub.mylibrary();

Methods

impl<'a, C, A> MylibraryMethods<'a, C, A>

pub fn bookshelves_clear_volumes(
&self,
shelf: &str
) -> MylibraryBookshelveClearVolumeCall<'a, C, A>

Create a builder to help you perform the following task:

Clears all volumes from a bookshelf.

Arguments

shelf - ID of bookshelf from which to remove a volume.

pub fn bookshelves_move_volume(
&self,
shelf: &str,
volume_id: &str,
volume_position: i32
) -> MylibraryBookshelveMoveVolumeCall<'a, C, A>

Create a builder to help you perform the following task:

Moves a volume within a bookshelf.

Arguments

shelf - ID of bookshelf with the volume.

volumeId - ID of volume to move.

volumePosition - Position on shelf to move the item (0 puts the item before the current first item, 1 puts it between the first and the second and so on.)

pub fn bookshelves_volumes_list(
&self,
shelf: &str
) -> MylibraryBookshelveVolumeListCall<'a, C, A>

Create a builder to help you perform the following task:

Gets volume information for volumes on a bookshelf.

Arguments

shelf - The bookshelf ID or name retrieve volumes for.

pub fn annotations_summary(
&self,
layer_ids: &Vec<String>,
volume_id: &str
) -> MylibraryAnnotationSummaryCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the summary of specified layers.

Arguments

layerIds - Array of layer IDs to get the summary for.

volumeId - Volume id to get the summary for.

pub fn annotations_delete(
&self,
annotation_id: &str
) -> MylibraryAnnotationDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an annotation.

Arguments

annotationId - The ID for the annotation to delete.

pub fn bookshelves_add_volume(
&self,
shelf: &str,
volume_id: &str
) -> MylibraryBookshelveAddVolumeCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a volume to a bookshelf.

Arguments

shelf - ID of bookshelf to which to add a volume.

volumeId - ID of volume to add.

pub fn annotations_insert(
&self,
request: Annotation
) -> MylibraryAnnotationInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts a new annotation.

Arguments

request - No description provided.

pub fn bookshelves_remove_volume(
&self,
shelf: &str,
volume_id: &str
) -> MylibraryBookshelveRemoveVolumeCall<'a, C, A>

Create a builder to help you perform the following task:

Removes a volume from a bookshelf.

Arguments

shelf - ID of bookshelf from which to remove a volume.

volumeId - ID of volume to remove.

pub fn annotations_list(&self) -> MylibraryAnnotationListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of annotations, possibly filtered.

pub fn annotations_update(
&self,
request: Annotation,
annotation_id: &str
) -> MylibraryAnnotationUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing annotation.

Arguments

request - No description provided.

annotationId - The ID for the annotation to update.

pub fn readingpositions_set_position(
&self,
volume_id: &str,
timestamp: &str,
position: &str
) -> MylibraryReadingpositionSetPositionCall<'a, C, A>

Create a builder to help you perform the following task:

Sets my reading position information for a volume.

Arguments

volumeId - ID of volume for which to update the reading position.

timestamp - RFC 3339 UTC format timestamp associated with this reading position.

position - Position string for the new volume reading position.

pub fn bookshelves_get(
&self,
shelf: &str
) -> MylibraryBookshelveGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves metadata for a specific bookshelf belonging to the authenticated user.

Arguments

shelf - ID of bookshelf to retrieve.

pub fn bookshelves_list(&self) -> MylibraryBookshelveListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of bookshelves belonging to the authenticated user.

pub fn readingpositions_get(
&self,
volume_id: &str
) -> MylibraryReadingpositionGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves my reading position information for a volume.

Arguments

volumeId - ID of volume for which to retrieve a reading position.