Struct google-books1::MylibraryMethodsBuilder

pub struct MylibraryMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on mylibrary resources. It is not used directly, but through the Books hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-books1" as books1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `annotations_delete(...)`, `annotations_insert(...)`, `annotations_list(...)`, `annotations_summary(...)`, `annotations_update(...)`, `bookshelves_add_volume(...)`, `bookshelves_clear_volumes(...)`, `bookshelves_get(...)`, `bookshelves_list(...)`, `bookshelves_move_volume(...)`, `bookshelves_remove_volume(...)`, `bookshelves_volumes_list(...)`, `readingpositions_get(...)` and `readingpositions_set_position(...)`
// to build up your call.
let rb = hub.mylibrary();

Methods

impl<'a, C, NC, A> MylibraryMethodsBuilder<'a, C, NC, A>

fn bookshelves_clear_volumes(&self, shelf: &str) -> MylibraryBookshelveClearVolumeCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Clears all volumes from a bookshelf.

fn bookshelves_move_volume(&self, shelf: &str, volume_id: &str, volume_position: i32) -> MylibraryBookshelveMoveVolumeCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Moves a volume within a bookshelf.

fn bookshelves_volumes_list(&self, shelf: &str) -> MylibraryBookshelveVolumeListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets volume information for volumes on a bookshelf.

fn annotations_summary(&self, layer_ids: &Vec<String>, volume_id: &str) -> MylibraryAnnotationSummaryCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the summary of specified layers.

fn annotations_delete(&self, annotation_id: &str) -> MylibraryAnnotationDeleteCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes an annotation.

fn bookshelves_add_volume(&self, shelf: &str, volume_id: &str) -> MylibraryBookshelveAddVolumeCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Adds a volume to a bookshelf.

fn annotations_insert(&self, request: &Annotation) -> MylibraryAnnotationInsertCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Inserts a new annotation.

fn bookshelves_remove_volume(&self, shelf: &str, volume_id: &str) -> MylibraryBookshelveRemoveVolumeCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Removes a volume from a bookshelf.

fn annotations_list(&self) -> MylibraryAnnotationListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves a list of annotations, possibly filtered.

fn annotations_update(&self, request: &Annotation, annotation_id: &str) -> MylibraryAnnotationUpdateCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing annotation.

fn readingpositions_set_position(&self, volume_id: &str, timestamp: &str, position: &str) -> MylibraryReadingpositionSetPositionCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sets my reading position information for a volume.

fn bookshelves_get(&self, shelf: &str) -> MylibraryBookshelveGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves metadata for a specific bookshelf belonging to the authenticated user.

fn bookshelves_list(&self) -> MylibraryBookshelveListCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves a list of bookshelves belonging to the authenticated user.

fn readingpositions_get(&self, volume_id: &str) -> MylibraryReadingpositionGetCallBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves my reading position information for a volume.