Struct google_books1::PromoofferMethods

pub struct PromoofferMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on promooffer resources. It is not used directly, but through the Books hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_books1 as books1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `accept(...)`, `dismiss(...)` and `get(...)`
// to build up your call.
let rb = hub.promooffer();

Methods

impl<'a, C, A> PromoofferMethods<'a, C, A>

fn accept(&self) -> PromoofferAcceptCall<'a, C, A>

Create a builder to help you perform the following task:

fn dismiss(&self) -> PromoofferDismisCall<'a, C, A>

Create a builder to help you perform the following task:

fn get(&self) -> PromoofferGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a list of promo offers available to the user