Struct google_books1::SeryMethods

pub struct SeryMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on sery resources. It is not used directly, but through the Books hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_books1 as books1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)` and `membership_get(...)`
// to build up your call.
let rb = hub.series();

Methods

impl<'a, C, A> SeryMethods<'a, C, A>

fn get(&self, series_id: &Vec<String>) -> SeryGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns Series metadata for the given series ids.

Arguments

series_id - String that identifies the series

fn membership_get(&self, series_id: &str) -> SeryMembershipGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns Series membership data given the series id.

Arguments

series_id - String that identifies the series