Struct google_books1::Volume

pub struct Volume {
    pub kind: Option<String>,
    pub access_info: Option<VolumeAccessInfo>,
    pub search_info: Option<VolumeSearchInfo>,
    pub sale_info: Option<VolumeSaleInfo>,
    pub etag: Option<String>,
    pub layer_info: Option<VolumeLayerInfo>,
    pub volume_info: Option<VolumeVolumeInfo>,
    pub recommended_info: Option<VolumeRecommendedInfo>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub user_info: Option<VolumeUserInfo>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

useruploaded list volumes (none)

recommended rate volumes (none)

list volumes (none)

associated list volumes (none)

get volumes (response)

recommended list volumes (none)

mybooks list volumes (none)

Fields

kind

Resource type for a volume. (In LITE projection.)

access_info

Any information about a volume related to reading or obtaining that volume text. This information can depend on country (books may be public domain in one country but not in another, e.g.).

search_info

Search result information related to this volume.

sale_info

Any information about a volume related to the eBookstore and/or purchaseability. This information can depend on the country where the request originates from (i.e. books may not be for sale in certain countries).

etag

Opaque identifier for a specific version of a volume resource. (In LITE projection)

layer_info

What layers exist in this volume and high level information about them.

volume_info

General volume information.

recommended_info

Recommendation related information for this volume.

id

Unique identifier for a volume. (In LITE projection.)

self_link

URL to this resource. (In LITE projection.)

user_info

User specific information related to this volume. (e.g. page this user last read or whether they purchased this book)