Struct google_books1::VolumeMethods

pub struct VolumeMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on volume resources. It is not used directly, but through the Books hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_books1 as books1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `associated_list(...)`, `get(...)`, `list(...)`, `mybooks_list(...)`, `recommended_list(...)`, `recommended_rate(...)` and `useruploaded_list(...)`
// to build up your call.
let rb = hub.volumes();

Methods

impl<'a, C, A> VolumeMethods<'a, C, A>

fn recommended_rate(&self, rating: &str, volume_id: &str) -> VolumeRecommendedRateCall<'a, C, A>

Create a builder to help you perform the following task:

Rate a recommended book for the current user.

Arguments

rating - Rating to be given to the volume.

volumeId - ID of the source volume.

fn mybooks_list(&self) -> VolumeMybookListCall<'a, C, A>

Create a builder to help you perform the following task:

Return a list of books in My Library.

fn list(&self, q: &str) -> VolumeListCall<'a, C, A>

Create a builder to help you perform the following task:

Performs a book search.

Arguments

q - Full-text search query string.

fn useruploaded_list(&self) -> VolumeUseruploadedListCall<'a, C, A>

Create a builder to help you perform the following task:

Return a list of books uploaded by the current user.

fn associated_list(&self, volume_id: &str) -> VolumeAssociatedListCall<'a, C, A>

Create a builder to help you perform the following task:

Return a list of associated books.

Arguments

volumeId - ID of the source volume.

fn get(&self, volume_id: &str) -> VolumeGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets volume information for a single volume.

Arguments

volumeId - ID of volume to retrieve.

fn recommended_list(&self) -> VolumeRecommendedListCall<'a, C, A>

Create a builder to help you perform the following task:

Return a list of recommended books for the current user.