Struct google_books1::VolumeSaleInfo

pub struct VolumeSaleInfo {
    pub country: Option<String>,
    pub retail_price: Option<VolumeSaleInfoRetailPrice>,
    pub is_ebook: Option<bool>,
    pub offers: Option<Vec<VolumeSaleInfoOffers>>,
    pub saleability: Option<String>,
    pub buy_link: Option<String>,
    pub on_sale_date: Option<String>,
    pub list_price: Option<VolumeSaleInfoListPrice>,
}

Any information about a volume related to the eBookstore and/or purchaseability. This information can depend on the country where the request originates from (i.e. books may not be for sale in certain countries).

This type is not used in any activity, and only used as part of another schema.

Fields

country

The two-letter ISO_3166-1 country code for which this sale information is valid. (In LITE projection.)

retail_price

The actual selling price of the book. This is the same as the suggested retail or list price unless there are offers or discounts on this volume. (In LITE projection.)

is_ebook

Whether or not this volume is an eBook (can be added to the My eBooks shelf).

offers

Offers available for this volume (sales and rentals).

saleability

Whether or not this book is available for sale or offered for free in the Google eBookstore for the country listed above. Possible values are FOR_SALE, FOR_RENTAL_ONLY, FOR_SALE_AND_RENTAL, FREE, NOT_FOR_SALE, or FOR_PREORDER.

buy_link

URL to purchase this volume on the Google Books site. (In LITE projection)

on_sale_date

The date on which this book is available for sale.

list_price

Suggested retail price. (In LITE projection.)