Struct google-books1::VolumeUserInfo

pub struct VolumeUserInfo {
    pub is_in_my_books: bool,
    pub rental_period: VolumeUserInfoRentalPeriod,
    pub updated: String,
    pub user_uploaded_volume_info: VolumeUserInfoUserUploadedVolumeInfo,
    pub rental_state: String,
    pub is_purchased: bool,
    pub reading_position: ReadingPosition,
    pub is_preordered: bool,
    pub copy: VolumeUserInfoCopy,
    pub review: Review,
    pub is_uploaded: bool,
}

User specific information related to this volume. (e.g. page this user last read or whether they purchased this book)

This type is not used in any activity, and only used as part of another schema.

Fields

is_in_my_books

Whether or not this volume is currently in "my books."

rental_period

Period during this book is/was a valid rental.

updated

Timestamp when this volume was last modified by a user action, such as a reading position update, volume purchase or writing a review. (RFC 3339 UTC date-time format).

user_uploaded_volume_info

no description provided

rental_state

Whether this book is an active or an expired rental.

is_purchased

Whether or not this volume was purchased by the authenticated user making the request. (In LITE projection.)

reading_position

The user's current reading position in the volume, if one is available. (In LITE projection.)

is_preordered

Whether or not this volume was pre-ordered by the authenticated user making the request. (In LITE projection.)

copy

Copy/Paste accounting information.

review

This user's review of this volume, if one exists.

is_uploaded

Whether or not this volume was user uploaded.