Struct google_books1::Volumeannotation

pub struct Volumeannotation {
    pub annotation_type: Option<String>,
    pub kind: Option<String>,
    pub deleted: Option<bool>,
    pub content_ranges: Option<VolumeannotationContentRanges>,
    pub updated: Option<String>,
    pub volume_id: Option<String>,
    pub annotation_data_id: Option<String>,
    pub annotation_data_link: Option<String>,
    pub page_ids: Option<Vec<String>>,
    pub layer_id: Option<String>,
    pub selected_text: Option<String>,
    pub data: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

volume annotations get layers (response)

Fields
annotation_type: Option<String>

The type of annotation this is.

kind: Option<String>

Resource Type

deleted: Option<bool>

Indicates that this annotation is deleted.

content_ranges: Option<VolumeannotationContentRanges>

The content ranges to identify the selected text.

updated: Option<String>

Timestamp for the last time this anntoation was updated. (RFC 3339 UTC date-time format).

volume_id: Option<String>

The Volume this annotation is for.

annotation_data_id: Option<String>

The annotation data id for this volume annotation.

annotation_data_link: Option<String>

Link to get data for this annotation.

page_ids: Option<Vec<String>>

Pages the annotation spans.

layer_id: Option<String>

The Layer this annotation is for.

selected_text: Option<String>

Excerpt from the volume.

data: Option<String>

Data for this annotation.

id: Option<String>

Unique id of this volume annotation.

self_link: Option<String>

URL to this resource.