Struct google_books1::Volumeseriesinfo

pub struct Volumeseriesinfo {
    pub book_display_number: Option<String>,
    pub kind: Option<String>,
    pub short_series_book_title: Option<String>,
    pub volume_series: Option<Vec<VolumeseriesinfoVolumeSeries>>,
}

There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Fields

book_display_number

The display number string. This should be used only for display purposes and the actual sequence should be inferred from the below orderNumber.

kind

Resource type.

short_series_book_title

Short book title in the context of the series.

volume_series

no description provided