Struct google_books1::LayerMethods

pub struct LayerMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on layer resources. It is not used directly, but through the Books hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_books1 as books1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use books1::Books;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Books::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `annotation_data_get(...)`, `annotation_data_list(...)`, `get(...)`, `list(...)`, `volume_annotations_get(...)` and `volume_annotations_list(...)`
// to build up your call.
let rb = hub.layers();

Methods

impl<'a, C, A> LayerMethods<'a, C, A>

fn annotation_data_get(&self, volume_id: &str, layer_id: &str, annotation_data_id: &str, content_version: &str) -> LayerAnnotationDataGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the annotation data.

Arguments

volumeId - The volume to retrieve annotations for.

layerId - The ID for the layer to get the annotations.

annotationDataId - The ID of the annotation data to retrieve.

contentVersion - The content version for the volume you are trying to retrieve.

fn volume_annotations_get(&self, volume_id: &str, layer_id: &str, annotation_id: &str) -> LayerVolumeAnnotationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the volume annotation.

Arguments

volumeId - The volume to retrieve annotations for.

layerId - The ID for the layer to get the annotations.

annotationId - The ID of the volume annotation to retrieve.

fn list(&self, volume_id: &str) -> LayerListCall<'a, C, A>

Create a builder to help you perform the following task:

List the layer summaries for a volume.

Arguments

volumeId - The volume to retrieve layers for.

fn get(&self, volume_id: &str, summary_id: &str) -> LayerGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the layer summary for a volume.

Arguments

volumeId - The volume to retrieve layers for.

summaryId - The ID for the layer to get the summary for.

fn volume_annotations_list(&self, volume_id: &str, layer_id: &str, content_version: &str) -> LayerVolumeAnnotationListCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the volume annotations for a volume and layer.

Arguments

volumeId - The volume to retrieve annotations for.

layerId - The ID for the layer to get the annotations.

contentVersion - The content version for the requested volume.

fn annotation_data_list(&self, volume_id: &str, layer_id: &str, content_version: &str) -> LayerAnnotationDataListCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the annotation data for a volume and layer.

Arguments

volumeId - The volume to retrieve annotation data for.

layerId - The ID for the layer to get the annotation data.

contentVersion - The content version for the requested volume.