Struct google_books1::VolumeAccessInfo

pub struct VolumeAccessInfo {
    pub web_reader_link: Option<String>,
    pub public_domain: Option<bool>,
    pub explicit_offline_license_management: Option<bool>,
    pub embeddable: Option<bool>,
    pub download_access: Option<DownloadAccessRestriction>,
    pub country: Option<String>,
    pub view_order_url: Option<String>,
    pub text_to_speech_permission: Option<String>,
    pub drive_imported_content_link: Option<String>,
    pub pdf: Option<VolumeAccessInfoPdf>,
    pub quote_sharing_allowed: Option<bool>,
    pub viewability: Option<String>,
    pub epub: Option<VolumeAccessInfoEpub>,
    pub access_view_status: Option<String>,
}

Any information about a volume related to reading or obtaining that volume text. This information can depend on country (books may be public domain in one country but not in another, e.g.).

This type is not used in any activity, and only used as part of another schema.

Fields

web_reader_link

URL to read this volume on the Google Books site. Link will not allow users to read non-viewable volumes.

public_domain

Whether or not this book is public domain in the country listed above.

explicit_offline_license_management

Whether this volume requires that the client explicitly request offline download license rather than have it done automatically when loading the content, if the client supports it.

embeddable

Whether this volume can be embedded in a viewport using the Embedded Viewer API.

download_access

Information about a volume's download license access restrictions.

country

The two-letter ISO_3166-1 country code for which this access information is valid. (In LITE projection.)

view_order_url

For ordered but not yet processed orders, we give a URL that can be used to go to the appropriate Google Wallet page.

text_to_speech_permission

Whether text-to-speech is permitted for this volume. Values can be ALLOWED, ALLOWED_FOR_ACCESSIBILITY, or NOT_ALLOWED.

drive_imported_content_link

URL to the Google Drive viewer if this volume is uploaded by the user by selecting the file from Google Drive.

pdf

Information about pdf content. (In LITE projection.)

quote_sharing_allowed

Whether quote sharing is allowed for this volume.

viewability

The read access of a volume. Possible values are PARTIAL, ALL_PAGES, NO_PAGES or UNKNOWN. This value depends on the country listed above. A value of PARTIAL means that the publisher has allowed some portion of the volume to be viewed publicly, without purchase. This can apply to eBooks as well as non-eBooks. Public domain books will always have a value of ALL_PAGES.

epub

Information about epub content. (In LITE projection.)

access_view_status

Combines the access and viewability of this volume into a single status field for this user. Values can be FULL_PURCHASED, FULL_PUBLIC_DOMAIN, SAMPLE or NONE. (In LITE projection.)