Struct google_books1::VolumeUserInfo

pub struct VolumeUserInfo {
    pub is_family_shared_from_user: Option<bool>,
    pub rental_period: Option<VolumeUserInfoRentalPeriod>,
    pub acquisition_type: Option<i32>,
    pub is_purchased: Option<bool>,
    pub updated: Option<String>,
    pub is_preordered: Option<bool>,
    pub copy: Option<VolumeUserInfoCopy>,
    pub is_family_sharing_disabled_by_fop: Option<bool>,
    pub is_uploaded: Option<bool>,
    pub acquired_time: Option<String>,
    pub is_in_my_books: Option<bool>,
    pub user_uploaded_volume_info: Option<VolumeUserInfoUserUploadedVolumeInfo>,
    pub family_sharing: Option<VolumeUserInfoFamilySharing>,
    pub rental_state: Option<String>,
    pub is_family_sharing_allowed: Option<bool>,
    pub reading_position: Option<ReadingPosition>,
    pub is_family_shared_to_user: Option<bool>,
    pub entitlement_type: Option<i32>,
    pub review: Option<Review>,
}

User specific information related to this volume. (e.g. page this user last read or whether they purchased this book)

This type is not used in any activity, and only used as part of another schema.

Fields
is_family_shared_from_user: Option<bool>

Whether or not the user shared this volume with the family.

rental_period: Option<VolumeUserInfoRentalPeriod>

Period during this book is/was a valid rental.

acquisition_type: Option<i32>

How this volume was acquired.

is_purchased: Option<bool>

Whether or not this volume was purchased by the authenticated user making the request. (In LITE projection.)

updated: Option<String>

Timestamp when this volume was last modified by a user action, such as a reading position update, volume purchase or writing a review. (RFC 3339 UTC date-time format).

is_preordered: Option<bool>

Whether or not this volume was pre-ordered by the authenticated user making the request. (In LITE projection.)

copy: Option<VolumeUserInfoCopy>

Copy/Paste accounting information.

is_family_sharing_disabled_by_fop: Option<bool>

Deprecated: Replaced by familySharing.

is_uploaded: Option<bool>

Whether or not this volume was user uploaded.

acquired_time: Option<String>

Timestamp when this volume was acquired by the user. (RFC 3339 UTC date-time format) Acquiring includes purchase, user upload, receiving family sharing, etc.

is_in_my_books: Option<bool>

Whether or not this volume is currently in "my books."

user_uploaded_volume_info: Option<VolumeUserInfoUserUploadedVolumeInfo>

no description provided

family_sharing: Option<VolumeUserInfoFamilySharing>

Information on the ability to share with the family.

rental_state: Option<String>

Whether this book is an active or an expired rental.

is_family_sharing_allowed: Option<bool>

Deprecated: Replaced by familySharing.

reading_position: Option<ReadingPosition>

The user's current reading position in the volume, if one is available. (In LITE projection.)

is_family_shared_to_user: Option<bool>

Whether or not the user received this volume through family sharing.

entitlement_type: Option<i32>

Whether this volume is purchased, sample, pd download etc.

review: Option<Review>

This user's review of this volume, if one exists.