Struct google_books1::VolumeVolumeInfo

pub struct VolumeVolumeInfo {
    pub subtitle: Option<String>,
    pub series_info: Option<Volumeseriesinfo>,
    pub page_count: Option<i32>,
    pub image_links: Option<VolumeVolumeInfoImageLinks>,
    pub ratings_count: Option<i32>,
    pub main_category: Option<String>,
    pub panelization_summary: Option<VolumeVolumeInfoPanelizationSummary>,
    pub authors: Option<Vec<String>>,
    pub categories: Option<Vec<String>>,
    pub description: Option<String>,
    pub publisher: Option<String>,
    pub dimensions: Option<VolumeVolumeInfoDimensions>,
    pub language: Option<String>,
    pub preview_link: Option<String>,
    pub maturity_rating: Option<String>,
    pub info_link: Option<String>,
    pub published_date: Option<String>,
    pub print_type: Option<String>,
    pub sample_page_count: Option<i32>,
    pub reading_modes: Option<String>,
    pub content_version: Option<String>,
    pub printed_page_count: Option<i32>,
    pub industry_identifiers: Option<Vec<VolumeVolumeInfoIndustryIdentifiers>>,
    pub title: Option<String>,
    pub average_rating: Option<f64>,
    pub allow_anon_logging: Option<bool>,
    pub canonical_volume_link: Option<String>,
}

General volume information.

This type is not used in any activity, and only used as part of another schema.

Fields
subtitle: Option<String>

Volume subtitle. (In LITE projection.)

series_info: Option<Volumeseriesinfo>

no description provided

page_count: Option<i32>

Total number of pages as per publisher metadata.

image_links: Option<VolumeVolumeInfoImageLinks>

A list of image links for all the sizes that are available. (In LITE projection.)

ratings_count: Option<i32>

The number of review ratings for this volume.

main_category: Option<String>

The main category to which this volume belongs. It will be the category from the categories list returned below that has the highest weight.

panelization_summary: Option<VolumeVolumeInfoPanelizationSummary>

A top-level summary of the panelization info in this volume.

authors: Option<Vec<String>>

The names of the authors and/or editors for this volume. (In LITE projection)

categories: Option<Vec<String>>

A list of subject categories, such as "Fiction", "Suspense", etc.

description: Option<String>

A synopsis of the volume. The text of the description is formatted in HTML and includes simple formatting elements, such as b, i, and br tags. (In LITE projection.)

publisher: Option<String>

Publisher of this volume. (In LITE projection.)

dimensions: Option<VolumeVolumeInfoDimensions>

Physical dimensions of this volume.

language: Option<String>

Best language for this volume (based on content). It is the two-letter ISO 639-1 code such as 'fr', 'en', etc.

preview_link: Option<String>

URL to preview this volume on the Google Books site.

maturity_rating: Option<String>

no description provided

info_link: Option<String>

URL to view information about this volume on the Google Books site. (In LITE projection)

published_date: Option<String>

Date of publication. (In LITE projection.)

print_type: Option<String>

Type of publication of this volume. Possible values are BOOK or MAGAZINE.

sample_page_count: Option<i32>

Total number of sample pages as per publisher metadata.

reading_modes: Option<String>

The reading modes available for this volume.

content_version: Option<String>

An identifier for the version of the volume content (text & images). (In LITE projection)

printed_page_count: Option<i32>

Total number of printed pages in generated pdf representation.

industry_identifiers: Option<Vec<VolumeVolumeInfoIndustryIdentifiers>>

Industry standard identifiers for this volume.

title: Option<String>

Volume title. (In LITE projection.)

average_rating: Option<f64>

The mean review rating for this volume. (min = 1.0, max = 5.0)

allow_anon_logging: Option<bool>

Whether anonymous logging should be allowed.

canonical_volume_link: Option<String>

Canonical URL for a volume. (In LITE projection.)